pollen_classification_ui/src/requirements.txt
numpy==2.0.1
Pillow==10.4.0
PyQt6==6.7.1
